+++ components/ADSList.module.css
@@ -1,0 +1,14 @@
+.listShell {
+    width:600px;
+    margin-top:60px;
+}
+
+.searchShell {
+    margin-bottom:20px;
+}
+
+.listShell table {
+    width:100%;
+    border-collapse:collapse;
+    
+}
